requirements.txt
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
tomli==2.0.1 ; python_version < "3.11"
